Move attributes and context to lib.rs
