+++ pyqir/src/lib.rs
@@ -14,7 +14,6 @@ extern crate llvm_sys_130 as llvm_sys;
 extern crate llvm_sys_140 as llvm_sys;
 
 mod builder;
-mod context;
 mod evaluator;
 mod instructions;
 mod module;
@@ -22,3 +21,75 @@ mod python;
 mod qis;
 mod types;
 mod values;
+
+use inkwell::attributes::AttributeLoc;
+use pyo3::prelude::*;
+use std::ops::Deref;
+
+/// The context owns global state needed by most LLVM objects.
+#[pyclass]
+#[derive(Eq, PartialEq)]
+pub(crate) struct Context(inkwell::context::Context);
+
+#[pymethods]
+impl Context {
+    #[new]
+    pub(crate) fn new() -> Self {
+        Self(inkwell::context::Context::create())
+    }
+}
+
+impl Deref for Context {
+    type Target = inkwell::context::Context;
+
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
+/// An attribute.
+#[pyclass(unsendable)]
+pub(crate) struct Attribute(pub(crate) inkwell::attributes::Attribute);
+
+#[pymethods]
+impl Attribute {
+    /// The value of this attribute as a string, or `None` if this is not a string attribute.
+    ///
+    /// :type: typing.Optional[str]
+    #[getter]
+    fn string_value(&self) -> Option<&str> {
+        if self.0.is_string() {
+            Some(
+                self.0
+                    .get_string_value()
+                    .to_str()
+                    .expect("Value is not valid UTF-8."),
+            )
+        } else {
+            None
+        }
+    }
+}
+
+/// The position of an attribute within a function declaration.
+#[pyclass]
+pub(crate) struct AttributeIndex(pub(crate) AttributeLoc);
+
+#[pymethods]
+impl AttributeIndex {
+    #[classattr]
+    const FUNCTION: Self = Self(AttributeLoc::Function);
+
+    #[classattr]
+    const RETURN: Self = Self(AttributeLoc::Return);
+
+    /// The attribute index for the nth parameter, starting from zero.
+    ///
+    /// :param int n: The parameter number.
+    /// :returns: The attribute index.
+    /// :rtype: AttributeIndex
+    #[staticmethod]
+    fn param(n: u32) -> Self {
+        Self(AttributeLoc::Param(n))
+    }
+}
